Add basic MCP server
