+++ x10/clients/onboarding/modules/base_module.py
@@ -5,18 +5,18 @@ from aiohttp import ClientTimeout
 from eth_account.messages import SignableMessage
 from eth_typing import ChecksumAddress
 
-from x10.config import Config
+from x10.core.client_config import ClientConfig
 from x10.signing.onboarding import SignMessageCallback
 from x10.utils.http import get_url
 
 
 class BaseModule:
-    __config: Config
+    __config: ClientConfig
     __account_address: ChecksumAddress
     __sign_message: SignMessageCallback
     __session: Optional[aiohttp.ClientSession]
 
-    def __init__(self, config: Config, *, account_address: ChecksumAddress, sign_message: SignMessageCallback):
+    def __init__(self, config: ClientConfig, *, account_address: ChecksumAddress, sign_message: SignMessageCallback):
         super().__init__()
 
         self.__config = config
@@ -27,7 +27,7 @@ class BaseModule:
     def _get_url(self, path: str, *, query: Optional[Dict] = None, **path_params) -> str:
         return get_url(f"{self.__config.endpoints.onboarding_url}{path}", query=query, **path_params)
 
-    def _get_config(self) -> Config:
+    def _get_config(self) -> ClientConfig:
         return self.__config
 
     def _get_account_address(self) -> ChecksumAddress:
