+++ x10/config.py
@@ -1,63 +1,16 @@
-from dataclasses import dataclass
 from decimal import Decimal
 
-
-@dataclass(kw_only=True, frozen=True)
-class StarknetDomain:
-    name: str
-    version: str
-    chain_id: str
-    revision: str
-
-
-@dataclass(kw_only=True, frozen=True)
-class DefaultsConfig:
-    market_price_slippage: Decimal
-    request_timeout_seconds: int
-
-
-@dataclass(kw_only=True, frozen=True)
-class SigningConfig:
-    signing_domain: str
-    starknet_domain: StarknetDomain
-
-
-@dataclass(kw_only=True, frozen=True)
-class EndpointsConfig:
-    """
-    Attributes:
-        chain_rpc_url (str): Field is deprecated and will be removed.
-        asset_operations_contract (str): Field is deprecated and will be removed.
-        collateral_asset_contract (str): Field is deprecated and will be removed.
-        collateral_asset_on_chain_id (str): Field is deprecated and will be removed.
-        collateral_decimals (int): Field is deprecated and will be removed.
-        collateral_asset_id (str): Field is deprecated and will be removed.
-    """
-
-    chain_rpc_url: str
-    api_base_url: str
-    stream_url: str
-    onboarding_url: str
-
-    asset_operations_contract: str
-    collateral_asset_contract: str
-    collateral_asset_on_chain_id: str
-    collateral_decimals: int
-    collateral_asset_id: str
-
-    vault_asset_name: str
-
-
-@dataclass(kw_only=True, frozen=True)
-class Config:
-    defaults: DefaultsConfig
-    signing: SigningConfig
-    endpoints: EndpointsConfig
-
+from core.client_config import (
+    ClientConfig,
+    DefaultsConfig,
+    EndpointsConfig,
+    SigningConfig,
+    StarknetDomain,
+)
 
 DEFAULTS = DefaultsConfig(market_price_slippage=Decimal("0.0075"), request_timeout_seconds=500)
 
-TESTNET_CONFIG = Config(
+TESTNET_CONFIG = ClientConfig(
     defaults=DEFAULTS,
     signing=SigningConfig(
         signing_domain="starknet.sepolia.extended.exchange",
@@ -78,7 +31,7 @@ TESTNET_CONFIG = Config(
 )
 
 
-MAINNET_CONFIG = Config(
+MAINNET_CONFIG = ClientConfig(
     defaults=DEFAULTS,
     signing=SigningConfig(
         signing_domain="extended.exchange",
