+++ x10/signing/transfer_object.py
@@ -3,7 +3,7 @@ from typing import List
 
 from fast_stark_crypto import get_transfer_msg_hash
 
-from x10.config import Config, StarknetDomain
+from x10.core.client_config import ClientConfig, StarknetDomain
 from x10.core.stark_account import StarkPerpetualAccount
 from x10.models.account import AccountModel
 from x10.models.base import SettlementSignatureModel
@@ -27,7 +27,7 @@ def create_transfer_object(
     to_vault: int,
     to_l2_key: int,
     amount: Decimal,
-    config: Config,
+    config: ClientConfig,
     stark_account: StarkPerpetualAccount,
     nonce: int | None = None,
 ) -> OnChainPerpetualTransferModel:
